test: split the suite along module boundaries
server.test.mjs was a single 2000-line file covering seven modules. It is
now four files that mirror the sources, plus test-helpers.ts for the
fixtures, loadServerAs() and the mock factories.

vitest scopes module mocks to the declaring file, so each test file repeats
the node:fs and node:fs/promises mocks; test-helpers.ts then picks up the
mocked copies through its own imports.

152 tests and 100% coverage unchanged.
